Factor out codeactions and formatting into their own modules
