+++ util/lsp/src/lib.rs
@@ -1,6 +1,8 @@
 mod capabilities;
+mod codeactions;
 mod conversion;
 mod diagnostics;
+mod format;
 mod hover;
 mod server;
